requirements.txt
annotated-types==0.7.0
annoy==1.17.3
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.11
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numpy==2.3.2
openai==1.98.0
orjson==3.11.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sqlalchemy-stubs==0.4
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zstandard==0.23.0
